Blocked Bloom filter and other improvements (#91)
* Cleaned up RollingHash. Added batch insertion to Bloom filter. Sort elements before querying Bloom filter. Added tests

* clean up leftovers from testing

* cleanup

* Renamed base and jump for clarity

* cleanup

* clarification

* cleanup

* Added TODO to Bloom filter

* scratch: batch check

* scratch

* scratch

* cleanup

* Added missing link

* rename Bloom filter tests

* clarifications. rearrange includes to match the others in tests/common

* minor

* more cache efficient

* minor

* Added comments to methods in bloom_filter.hpp

* Added comments to methods in kmer_bloom_filter.hpp

* early stop in BloomFilter::check

* minor

* make sure accesses to the filter are aligned to 512-bit blocks

* rename variable

* rename variables. update comments

* stop hogging xmm registers

* revert changes to CMake

* Added asserts

* optimizations

* optimization in ColumnCompressed

* Optimizations in alignment code. Implemented score updating with AVX2 instructions (#103)

* change alignment parameters

* Cleanup. Moved underflow checks to aligner init

* Update scores in aligner with AVX2 instructions

Removed DPTable::Step

more SIMD

Used aligned_allocator for update vectors

Cleanup and added comments

* Move compute_delete_scores to its own function

* assert fix

* fix max_pos

* Update help message with new params

* optimization

* minor

* Moved AVX2 code to helper functions. Cleanup

* minor

* Added missing include in test_helpers.hpp

* removed leftover comment

* cleanup

* Switch to faster method for restricting values to index range

* Update metagraph/src/graph/succinct/dbg_succinct.cpp


* Update metagraph/src/graph/succinct/dbg_succinct.cpp


* Cleanup

* Use the most significant bit to check if the hash is valid or invalid

* Use least significant bit to check if hash is present. Store hashes in AlignedVector

* minor change: fixed compilation on MacOS

* Removed hash restrictions on BloomFilter. Invalid k-mers are now internally hashed as npos in kmer_bloom_filter

* change buffering strategy in KmerBloomFilter::add_sequences

* clean up restrict_to

* minor

* Update metagraph/src/kmer/kmer_bloom_filter.hpp


* cleanup

* clean up rolling hash

* Added BloomFilter::check which uses a callback to report indices of set bits

* clarification

* Added call_valid_kmers. Fixed assert in add_sequences

+++ metagraph/src/graph/succinct/dbg_succinct.hpp
@@ -134,10 +134,10 @@ class DBGSuccinct : public DeBruijnGraph {
     node_index boss_to_kmer_index(uint64_t boss_index) const;
 
     void initialize_bloom_filter_from_fpr(double false_positive_rate,
-                                          size_t max_num_hash_functions = -1);
+                                          uint32_t max_num_hash_functions = -1);
 
     void initialize_bloom_filter(double bits_per_kmer,
-                                 size_t max_num_hash_functions = -1);
+                                 uint32_t max_num_hash_functions = -1);
 
     const KmerBloomFilter<>* get_bloom_filter() const { return bloom_filter_.get(); }
